test(e2e): hoist util and environment from wallet tests dir
