+++ packages/e2e/test/wallet/SingleAddressWallet/pouchDbWalletStores.test.ts
@@ -1,9 +1,9 @@
-import { env } from '../environment';
+import { env } from '../../environment';
 import { filter, firstValueFrom } from 'rxjs';
 import { getWallet } from '../../../src/factories';
 import { logger } from '@cardano-sdk/util-dev';
 import { storage } from '@cardano-sdk/wallet';
-import { waitForWalletStateSettle, walletReady } from '../util';
+import { waitForWalletStateSettle, walletReady } from '../../util';
 import delay from 'delay';
 
 describe('SingleAddressWallet/pouchDbWalletStores', () => {
