+++ packages/e2e/test/wallet/SingleAddressWallet/txChainHistory.test.ts
@@ -1,12 +1,12 @@
 /* eslint-disable max-len */
 import { SingleAddressWallet, buildTx } from '@cardano-sdk/wallet';
 import { assertTxIsValid } from '../../../../wallet/test/util';
-import { env } from '../environment';
+import { env } from '../../environment';
 import { filter, firstValueFrom, map, take } from 'rxjs';
 import { getWallet } from '../../../src';
 import { isNotNil } from '@cardano-sdk/util';
 import { logger } from '@cardano-sdk/util-dev';
-import { normalizeTxBody } from '../util';
+import { normalizeTxBody } from '../../util';
 
 describe('tx chain history', () => {
   let wallet: SingleAddressWallet;
